cosmetic improvement

+++ NumericUtil.cpp
@@ -30,6 +30,7 @@ limitations under the License.
 #include "SqMatrixGeneric.h"
 #include "PolynomialGeneric.h"
 
+#include <cstddef>
 #include <complex>
 #include <limits>
 
@@ -269,7 +270,7 @@ namespace math
  * @return base^n 
  */
 template<class T>
-T math::NumericUtil<T>::power(const T& base, unsigned int n)
+T math::NumericUtil<T>::power(const T& base, size_t n)
 {
     /*
      * "Exponentiation by squaring" algorithm will be applied.
@@ -302,11 +303,11 @@ T math::NumericUtil<T>::power(const T& base, unsigned int n)
     
     // Obtain coefficients ai from the exponent's binary form.
     // Note: "i>>=1" is a bitwise equivalent bitwise equivalent of "i/=2"
-    for ( unsigned int i=n; i>0; i>>=1 )
+    for ( size_t i=n; i>0; i>>=1 )
     {
         // Check the coefficient ai (no need to multiply retVal by 1 if ai is 0)
         // Note: "i&1" is a bitwise equivalent of "i%2"
-        if ( 0!=(i & static_cast<unsigned int>(1) ) )
+        if ( 0!=(i & static_cast<size_t>(1) ) )
         {
             retVal *= factor;
         }
